Move memory intensive test case to separate location
