+++ scripts/onnx/test.sh
@@ -57,7 +57,7 @@ if [[ "${SHARD_NUMBER}" == "1" ]]; then
   # These exclusions are for tests that take a long time / a lot of GPU
   # memory to run; they should be passing (and you will test them if you
   # run them locally
-  pytest "${args[@]}" \
+  pytest "${args[@]}" "${args_parallel[@]}" \
     --ignore "$top_dir/test/onnx/test_pytorch_onnx_onnxruntime.py" \
     --ignore "$top_dir/test/onnx/test_custom_ops.py" \
     --ignore "$top_dir/test/onnx/test_models_onnxruntime.py" \
@@ -70,8 +70,7 @@ if [[ "${SHARD_NUMBER}" == "1" ]]; then
 
   # Tests that cannot run in parallel.
   pytest "${args[@]}" \
-    "$top_dir/test/onnx/test_pytorch_onnx_caffe2.py" \
-    "$top_dir/test/onnx/test_pytorch_onnx_caffe2_quantized.py"
+    "$top_dir/test/onnx/test_onnx_export.py"
 
   pytest "${args[@]}" "${args_parallel[@]}" \
     "$top_dir/test/onnx/test_pytorch_onnx_onnxruntime.py::TestONNXRuntime_opset7" \
@@ -81,7 +80,8 @@ if [[ "${SHARD_NUMBER}" == "1" ]]; then
     "$top_dir/test/onnx/test_models_onnxruntime.py" \
     "$top_dir/test/onnx/test_utility_funs.py" \
     "$top_dir/test/onnx/test_pytorch_onnx_shape_inference.py" \
-    "$top_dir/test/onnx/test_onnx_export.py"
+    "$top_dir/test/onnx/test_pytorch_onnx_caffe2.py" \
+    "$top_dir/test/onnx/test_pytorch_onnx_caffe2_quantized.py"
 fi
 
 if [[ "${SHARD_NUMBER}" == "2" ]]; then
