+++ test/onnx/test_pytorch_onnx_onnxruntime.py
@@ -269,80 +269,6 @@ class _TestONNXRuntime:
         if not is_script:
             _run_test(model, tracing_remained_onnx_input_idx)
 
-    @skipIfUnsupportedMinOpsetVersion(
-        9
-    )  # Because external data format was released with Opset 9.
-    def test_embedding_model_with_external_data(self):
-        class LargeModel(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                dim = 15
-                n = 4 * 100
-                self.emb = torch.nn.Embedding(n, dim)
-                self.lin1 = torch.nn.Linear(dim, 1)
-                self.seq = torch.nn.Sequential(
-                    self.emb,
-                    self.lin1,
-                )
-
-            def forward(self, input):
-                return self.seq(input)
-
-        model = LargeModel()
-        x = torch.tensor([2], dtype=torch.long)
-        run_model_test_with_external_data(self, model, x)
-
-    @skipIfUnsupportedMinOpsetVersion(
-        9
-    )  # Because external data format was released with Opset 9.
-    def test_large_model_with_external_data(self):
-        class LargeModel(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                dim = 5
-                n = 40 * 4 * 10**6
-                self.emb = torch.nn.Embedding(n, dim)
-                self.lin1 = torch.nn.Linear(dim, 1)
-                self.seq = torch.nn.Sequential(
-                    self.emb,
-                    self.lin1,
-                )
-
-            def forward(self, input):
-                return self.seq(input)
-
-        x = torch.tensor([2], dtype=torch.long)
-        run_model_test_with_external_data(self, LargeModel(), x)
-
-    @skipIfUnsupportedMinOpsetVersion(
-        9
-    )  # Because external data format was released with Opset 9.
-    def test_large_model_with_non_str_file(self):
-        class LargeModel(torch.nn.Module):
-            def __init__(self):
-                super().__init__()
-                dim = 5
-                n = 40 * 4 * 10**6
-                self.emb = torch.nn.Embedding(n, dim)
-                self.lin1 = torch.nn.Linear(dim, 1)
-                self.seq = torch.nn.Sequential(
-                    self.emb,
-                    self.lin1,
-                )
-
-            def forward(self, input):
-                return self.seq(input)
-
-        x = torch.tensor([2], dtype=torch.long)
-        f = io.BytesIO()
-        err_msg = (
-            "The serialized model is larger than the 2GiB limit imposed by the protobuf library. "
-            "Therefore the output file must be a file path, so that the ONNX external data can be written to "
-            "the same directory. Please specify the output file name."
-        )
-        with self.assertRaisesRegex(RuntimeError, err_msg):
-            torch.onnx.export(LargeModel(), x, f)
-
     def test_fuse_conv_bn1d(self):
         class Fuse(torch.nn.Module):
             def __init__(self):
